Fix problem with SD printing; added M200, M202, M500 for runtime config

+++ R2C2-RepRap_firmware/RepRap/Mendel/gcode_parse.h
@@ -75,7 +75,7 @@ typedef struct {
 	uint8_t					option_inches				:1;
 
 	uint8_t						G;
-	uint8_t						M;
+	uint16_t				  M;
 	TARGET						target;
 
 	int16_t						S;
